+++ src/resolvers/User/likes/resolver.ts
@@ -1,9 +1,9 @@
 import { UserResolvers } from "../../graphql.js";
 import { ResolverDeps } from "../../index.js";
 import { MylistModel } from "../../Mylist/model.js";
-import { get } from "./get.js";
+import { get } from "./prisma.js";
 
-export const resolveUserLikes = ({ prisma }: Pick<ResolverDeps, "prisma">) =>
+export const resolverUserLikes = ({ prisma }: Pick<ResolverDeps, "prisma">) =>
   (async ({ id: userId }, _args, { user }) => {
     const result = await get(prisma, { holderId: userId, authUserId: user?.id || null });
